+++ app/lib/__init__.py
@@ -18,7 +18,7 @@ debug = conf.getboolean('Logging', 'debug')
 logger = logging.getLogger("lib")
 
 
-# TODO: Move to textHandling.py, update references and retest.
+# TODO: Move to text_handling.py, update references and retest.
 def flattenText(text, replacement=u" "):
     r"""
     Flatten a string from multi-line to a single line, using a specified
