+++ app/lib/query/tweets/categories.py
@@ -55,7 +55,7 @@ def printUnassignedProfiles():
     TODO: Add filters such as top N recently created profiles or most
     followers. And find a way to make this more useful, considering that
     the influencer category and a specific influencer category could be assigned
-    on fetchProfiles.py running, but it has to be clear that industry is
+    on fetch_profiles.py running, but it has to be clear that industry is
     assigned yet.
 
     @return: None
