+++ app/lib/text_handling.py
