+++ app/models/__init__.py
@@ -10,11 +10,11 @@ file is run, which causes a conflict.
 from .places import __all__ as placesModel
 from .trends import __all__ as trendsModel
 from .tweets import __all__ as tweetsModel
-from .cronJobs import __all__ as cronJobsModel
+from .cron_jobs import __all__ as cronJobsModel
 __all__ = placesModel + trendsModel + tweetsModel + cronJobsModel
 
 # Make table objects available on models module.
 from .places import *
 from .trends import *
 from .tweets import *
-from .cronJobs import *
+from .cron_jobs import *
