+++ app/utils/insert/fetch_profiles.py
