+++ app/utils/reporting/profile_to_word_network.py
